refactor(ci): update workflow, script, and tests to use latest action version for integration tests
