--- eox_tenant/test/tutor/pytest.ini
@@ -1,7 +1,0 @@
-[pytest]
-python_files = integration_test_*.py
-filterwarnings =
-    default
-    # We ignore every warning while we actually get the testing infrastructure
-    # running for different version of tutor in gh actions
-    ignore:
